requirements.txt
aniso8601==9.0.1
async-generator==1.10
attrs==21.4.0
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
distlib==0.3.4
exceptiongroup==1.0.4
filelock==3.7.1
Flask==1.1.2
Flask-JWT-Extended==4.3.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==1.0.1
jsonschema==4.5.1
MarkupSafe==1.1.1
marshmallow==3.14.1
outcome==1.2.0
packaging==22.0
platformdirs==2.5.2
PyMySQL==0.9.3
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzbar==0.1.9
rfc3986==1.5.0
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.37
SQLAlchemy-Utils==0.38.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.2.0
urllib3==1.26.9
validators==0.20.0
virtualenv==20.15.0
Werkzeug==0.16.1
wsproto==1.2.0
zipp==3.8.0
